+++ pcredz_ultimate/__init__.py
@@ -1,3 +1,3 @@
 """PCredz Ultimate passive network credential auditing package."""
 
-__version__ = '3.1.0'
+__version__ = '3.1.1'